Extracts weather functions into Weather module & rename WeatherClient
