+++ src/js/weather.js
@@ -1,166 +1,48 @@
-var Weather = {
-    Utils: {},
-    apiKey: ''
+import Icons from './icons';
+import WeatherClient from './weather-client';
+
+import Utilities from './utilities';
+const { delay, toTitleCase } = Utilities;
+
+function updateWeatherInformation(current){
+    const weatherSelector = document.getElementById('weather');
+    const iconSelector = weatherSelector.querySelector('.icon');
+
+    const temperature = Math.floor(current.temperature());
+    const weatherCond = toTitleCase(current.conditions());
+
+    const icon = Icons.getIcon(current.icon());
+
+    const iconRegex = /(^|\s)i-\S+/g;
+
+    weatherSelector.querySelector('.temp').innerHTML = temperature;
+
+    if(iconSelector.className.match(iconRegex)){
+        iconSelector.className = iconSelector.className.replace(iconRegex, '');
+    }
+    iconSelector.classList.add(icon.iconClass);
+
+    Icons.changeFavicon(`images/icons/${icon.iconName}.ico`);
+
+    weatherSelector.querySelector('.description').innerHTML = weatherCond;
+
+    document.querySelector('title').innerHTML = `${temperature}° ${weatherCond}`;
+}
+
+function displayWeather(city, cycle_duration = 30000){
+    delay(cycle_duration).then(() => {
+        WeatherClient.getCurrent(city, updateWeatherInformation);
+
+        displayWeather(city, cycle_duration);
+    });
+}
+
+const weather = {
+    initialize: function(weather_api_key, city, cycle_duration){
+        WeatherClient.apiKey = weather_api_key;
+
+        displayWeather(city, cycle_duration);
+    }
 };
 
-Weather.VERSION = '0.0.4';
-
-var jsonp = Weather.Utils.jsonp = function (uri){
-    return new Promise(function(resolve, reject){
-        var id = '_' + Math.round(10000 * Math.random());
-        var callbackName = 'jsonp_callback_' + id;
-        var el = (document.getElementsByTagName('head')[0] || document.body || document.documentElement);
-        var script = document.createElement('script');
-        var src = uri + '&callback=' + callbackName;
-
-        window[callbackName] = function(data){
-            delete window[callbackName];
-            var ele = document.getElementById(id);
-            ele.parentNode.removeChild(ele);
-            resolve(data);
-        };
-
-        script.src = src;
-        script.id = id;
-        script.addEventListener('error', reject);
-        el.appendChild(script);
-    } );
-};
-
-Weather.kelvinToFahrenheit = function (value) {
-    return (this.kelvinToCelsius(value) * 1.8) + 32;
-};
-
-Weather.kelvinToCelsius = function (value) {
-    return value - 273.15;
-};
-
-Weather.getCurrent = function (city, callback) {
-    var url = 'http://api.openweathermap.org/data/2.5/weather?q=' + encodeURIComponent(city) + '&units=imperial&&APPID=' + Weather.apiKey;
-
-    return this._getJSON(url, function (data) {
-        callback(new Weather.Current(data));
-    } );
-};
-
-Weather.getForecast = function (city, callback) {
-    var url = 'http://openweathermap.org/data/2.1/forecast/city?q=' + encodeURIComponent(city) + '&cnt=1';
-
-    return this._getJSON(url, function (data) {
-        callback(new Weather.Forecast(data));
-    } );
-};
-
-Weather._getJSON = function(url, callback) {
-    jsonp(url).then(callback);
-};
-
-var maxBy = Weather.Utils.maxBy = function (list, iterator) {
-    var max;
-    var f = function (memo, d) {
-        var val = iterator(d);
-
-        if (memo === null || val > max) {
-            max = val;
-            memo = d;
-        }
-
-        return memo;
-    };
-
-    return list.reduce(f, null);
-};
-
-var minBy = Weather.Utils.minBy = function (list, iterator) {
-    var min;
-    var f = function (memo, d) {
-        var val = iterator(d);
-
-        if (memo === null || val < min) {
-            min = val;
-            memo = d;
-        }
-
-        return memo;
-    };
-
-    return list.reduce(f, null);
-};
-
-var isOnDate = Weather.Utils.isOnDate = function (a, b) {
-    var sameYear = a.getYear() === b.getYear();
-    var sameMonth = a.getMonth() === b.getMonth();
-    var sameDate = a.getDate() === b.getDate();
-
-    return sameYear && sameMonth && sameDate;
-};
-
-Weather.Forecast = function (data) {
-    this.data = data;
-};
-
-Weather.Forecast.prototype.startAt = function () {
-    return new Date(minBy(this.data.list, function (d) { return d.dt; }).dt * 1000);
-};
-
-Weather.Forecast.prototype.endAt = function () {
-    return new Date(maxBy(this.data.list, function (d) { return d.dt; }).dt * 1000);
-};
-
-Weather.Forecast.prototype.day = function (date) {
-    return new Weather.Forecast(this._filter(date));
-};
-
-Weather.Forecast.prototype.low = function () {
-    if (this.data.list.length === 0) return;
-
-    var output = minBy(this.data.list, function (item) {
-        return item.main.temp_min;
-    } );
-
-    return output.main.temp_min;
-};
-
-Weather.Forecast.prototype.high = function () {
-    if (this.data.list.length === 0) return;
-
-    var output = maxBy( this.data.list, function (item) {
-        return item.main.temp_max;
-    } );
-
-    return output.main.temp_max;
-};
-
-Weather.Forecast.prototype._filter = function (date) {
-    return {
-        list: this.data.list.filter(function (range) {
-            var dateTimestamp = (range.dt * 1000);
-
-            if (isOnDate(new Date(dateTimestamp), date)) {
-                return range;
-            }
-        })
-    };
-};
-
-Weather.Current = function (data) {
-    //console.log(data)
-    this.data = data;
-};
-
-Weather.Current.prototype.temperature = function () {
-    return this.data.main.temp;
-};
-
-Weather.Current.prototype.conditions = function () {
-    return this.data.weather[0].description;
-};
-
-Weather.Current.prototype.icon = function () {
-    return {
-        code: this.data.weather[0].icon,
-        id: this.data.weather[0].id
-    };
-};
-
-export default Weather;
+export default weather;
